Segregate calculation steps
