+++ source/databricks/package/steps/total_production_per_ga.py
@@ -1,0 +1,140 @@
+# Copyright 2020 Energinet DataHub A/S
+#
+# Licensed under the Apache License, Version 2.0 (the "License2");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from pyspark.sql import DataFrame
+
+from pyspark.sql.functions import (
+    array,
+    array_contains,
+    lit,
+    col,
+    collect_set,
+    row_number,
+    expr,
+    when,
+    explode,
+    sum,
+)
+from pyspark.sql.types import (
+    DecimalType,
+)
+from pyspark.sql.window import Window
+from package.codelists import (
+    TimeSeriesQuality,
+    MeteringPointResolution,
+)
+
+from package.db_logging import debug
+from decimal import Decimal
+
+
+def get_total_production_per_ga_df(
+    enriched_time_series_points_df: DataFrame,
+) -> DataFrame:
+    # Total production in batch grid areas with quarterly resolution per grid area
+    result_df = (
+        enriched_time_series_points_df.withColumn(
+            "quarter_times",
+            when(
+                col("Resolution") == MeteringPointResolution.hour.value,
+                array(
+                    col("time"),
+                    col("time") + expr("INTERVAL 15 minutes"),
+                    col("time") + expr("INTERVAL 30 minutes"),
+                    col("time") + expr("INTERVAL 45 minutes"),
+                ),
+            ).when(
+                col("Resolution") == MeteringPointResolution.quarterly.value,
+                array(col("time")),
+            ),
+        )
+        .select(
+            enriched_time_series_points_df["*"],
+            explode("quarter_times").alias("quarter_time"),
+        )
+        .withColumn("Quantity", col("Quantity").cast(DecimalType(18, 6)))
+        .withColumn(
+            "quarter_quantity",
+            when(
+                col("Resolution") == MeteringPointResolution.hour.value,
+                col("Quantity") / 4,
+            ).when(
+                col("Resolution") == MeteringPointResolution.quarterly.value,
+                col("Quantity"),
+            ),
+        )
+        .groupBy("GridAreaCode", "quarter_time")
+        .agg(sum("quarter_quantity"), collect_set("Quality"))
+        .withColumn(
+            "Quality",
+            when(
+                array_contains(
+                    col("collect_set(Quality)"), lit(TimeSeriesQuality.missing.value)
+                ),
+                lit(TimeSeriesQuality.missing.value),
+            )
+            .when(
+                array_contains(
+                    col("collect_set(Quality)"),
+                    lit(TimeSeriesQuality.estimated.value),
+                ),
+                lit(TimeSeriesQuality.estimated.value),
+            )
+            .when(
+                array_contains(
+                    col("collect_set(Quality)"),
+                    lit(TimeSeriesQuality.measured.value),
+                ),
+                lit(TimeSeriesQuality.measured.value),
+            ),
+        )
+        .withColumnRenamed("Quality", "quality")
+    )
+
+    debug(
+        "Pre-result split into quarter times",
+        result_df.orderBy(col("GridAreaCode"), col("quarter_time")),
+    )
+
+    window = Window.partitionBy("GridAreaCode").orderBy(col("quarter_time"))
+
+    # Points may be missing in result time series if all metering points are missing a point at a certain moment.
+    # According to PO and SME we can for now assume that full time series have been submitted for the processes/tests in question.
+    result_df = (
+        result_df.withColumn("position", row_number().over(window))
+        .withColumnRenamed("sum(quarter_quantity)", "Quantity")
+        .withColumn(
+            "Quantity",
+            when(col("Quantity").isNull(), Decimal("0.000")).otherwise(col("Quantity")),
+        )
+        .withColumn(
+            "quality",
+            when(col("quality").isNull(), TimeSeriesQuality.missing.value).otherwise(
+                col("quality")
+            ),
+        )
+        .select(
+            "GridAreaCode",
+            col("Quantity").cast(DecimalType(18, 3)),
+            col("quality"),
+            "position",
+            "quarter_time",
+        )
+    )
+
+    debug(
+        "Balance fixing total production result",
+        result_df.orderBy(col("GridAreaCode"), col("position")),
+    )
+    return result_df
